layers/shared/requirements.txt
PyYAML>=6.0
requests
pydantic>=2,<3